REFACTOR: Todos DTO 폴더 구조 변경#30
