+++ src/main/java/com/ureca/fitlog/todos/dto/response/TodoDoneResponseDTO.java
@@ -1,4 +1,4 @@
-package com.ureca.fitlog.todos.dto;
+package com.ureca.fitlog.todos.dto.response;
 
 import lombok.*;
